Unit tests should be in the same file
