+++ vm/src/tests/mod.rs
@@ -14,5 +14,4 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-mod decoder;
 mod executor_tests;
